requirements.txt
python-nmap==0.7.1
jinja2==3.1.2
weasyprint==59.0
pyyaml==6.0.1
colorama==0.4.6
netifaces==0.11.0
paramiko==3.4.0
pymysql==1.1.0
